Move errorutils to the API
