+++ .bldr.toml
@@ -377,6 +377,7 @@ paths = [
   "api/external/event_feed/*",
   "api/external/infra_proxy/*",
   "api/external/ingest/*",
+  "api/external/lib/*",
   "api/external/nodes/*",
   "api/external/secrets/*",
   "api/interservice/authn/*",
@@ -405,7 +406,6 @@ paths = [
   "components/trial-license-service/*",
   "lib/authz/*",
   "lib/cereal/*",
-  "lib/errorutils/*",
   "lib/grpc/*",
   "lib/io/*",
   "lib/license/*",
@@ -495,6 +495,7 @@ paths = [
   "api/config/shared/*",
   "api/external/common/*",
   "api/external/ingest/*",
+  "api/external/lib/*",
   "api/external/secrets/*",
   "api/interservice/authn/*",
   "api/interservice/authz/*",
@@ -515,7 +516,6 @@ paths = [
   "lib/datalifecycle/*",
   "lib/db/*",
   "lib/elasticutil/*",
-  "lib/errorutils/*",
   "lib/grpc/*",
   "lib/io/*",
   "lib/logger/*",
@@ -1284,12 +1284,12 @@ paths = [
   "api/external/applications/*",
   "api/external/common/*",
   "api/external/habitat/*",
+  "api/external/lib/*",
   "api/interservice/cereal/*",
   "components/automate-gateway/api/iam/*",
   "components/automate-grpc/*",
   "lib/cereal/*",
   "lib/db/*",
-  "lib/errorutils/*",
   "lib/grpc/*",
   "lib/io/*",
   "lib/logger/*",
@@ -1398,6 +1398,7 @@ paths = [
   "api/config/platform/*",
   "api/config/shared/*",
   "api/external/common/*",
+  "api/external/lib/*",
   "api/external/secrets/*",
   "api/interservice/authz/*",
   "api/interservice/cereal/*",
@@ -1412,7 +1413,6 @@ paths = [
   "lib/authz/*",
   "lib/cereal/*",
   "lib/db/*",
-  "lib/errorutils/*",
   "lib/grpc/*",
   "lib/io/*",
   "lib/logger/*",
@@ -1721,11 +1721,11 @@ paths = [
   "api/config/platform/*",
   "api/config/shared/*",
   "api/external/common/*",
+  "api/external/lib/*",
   "api/external/secrets/*",
   "components/automate-gateway/api/iam/*",
   "components/automate-grpc/*",
   "lib/db/*",
-  "lib/errorutils/*",
   "lib/grpc/*",
   "lib/io/*",
   "lib/logger/*",
@@ -2024,6 +2024,7 @@ paths = [
   "api/external/cfgmgmt/*",
   "api/external/common/*",
   "api/external/data_feed/*",
+  "api/external/lib/*",
   "api/external/secrets/*",
   "api/interservice/authz/*",
   "api/interservice/cereal/*",
@@ -2036,7 +2037,6 @@ paths = [
   "lib/authz/*",
   "lib/cereal/*",
   "lib/db/*",
-  "lib/errorutils/*",
   "lib/grpc/*",
   "lib/io/*",
   "lib/logger/*",
@@ -2228,13 +2228,13 @@ paths = [
   "api/config/platform/*",
   "api/config/shared/*",
   "api/external/common/*",
+  "api/external/lib/*",
   "api/external/secrets/*",
   "api/interservice/authz/*",
   "api/interservice/infra_proxy/*",
   "components/automate-gateway/api/iam/*",
   "components/automate-grpc/*",
   "lib/db/*",
-  "lib/errorutils/*",
   "lib/grpc/*",
   "lib/io/*",
   "lib/logger/*",
