+++ components/compliance-service/api/stats/server/server.go
@@ -9,7 +9,7 @@ import (
 
 	"github.com/chef/automate/api/interservice/compliance/stats"
 	"github.com/chef/automate/components/compliance-service/reporting/relaxting"
-	"github.com/chef/automate/lib/errorutils"
+	"github.com/chef/automate/api/external/lib/errorutils"
 )
 
 // Server implementation for stats
