+++ components/compliance-service/dao/pgdb/filtering.go
@@ -7,7 +7,7 @@ import (
 
 	"github.com/chef/automate/api/interservice/compliance/common"
 	"github.com/chef/automate/components/compliance-service/utils"
-	"github.com/chef/automate/lib/errorutils"
+	"github.com/chef/automate/api/external/lib/errorutils"
 	"github.com/chef/automate/lib/pgutils"
 
 	"github.com/pkg/errors"
