+++ components/compliance-service/reporting/relaxting/nodes.go
@@ -13,7 +13,7 @@ import (
 
 	reportingapi "github.com/chef/automate/api/interservice/compliance/reporting"
 	"github.com/chef/automate/components/compliance-service/reporting"
-	"github.com/chef/automate/lib/errorutils"
+	"github.com/chef/automate/api/external/lib/errorutils"
 )
 
 type ProfileSource struct {
