+++ components/nodemanager-service/managers/awsec2/awsec2.go
@@ -29,7 +29,7 @@ import (
 	"github.com/chef/automate/components/compliance-service/inspec-agent/types"
 	"github.com/chef/automate/components/compliance-service/utils/pool"
 	"github.com/chef/automate/components/nodemanager-service/pgdb"
-	"github.com/chef/automate/lib/errorutils"
+	"github.com/chef/automate/api/external/lib/errorutils"
 	"github.com/chef/automate/lib/stringutils"
 )
