+++ components/nodemanager-service/managers/azure/azure.go
@@ -23,7 +23,7 @@ import (
 	"github.com/chef/automate/components/compliance-service/inspec-agent/types"
 	"github.com/chef/automate/components/compliance-service/utils/pool"
 	"github.com/chef/automate/components/nodemanager-service/pgdb"
-	"github.com/chef/automate/lib/errorutils"
+	"github.com/chef/automate/api/external/lib/errorutils"
 	"github.com/chef/automate/lib/stringutils"
 )
